Move nftables rule conversion functions to separate file
...to reduce file size.
