+++ pkg/packetfilter/nftables/nftables.go
@@ -19,16 +19,12 @@ limitations under the License.
 package nftables
 
 import (
-	"bytes"
 	"context"
 	"slices"
-	"strconv"
-	"strings"
 
 	"github.com/pkg/errors"
 	"github.com/submariner-io/admiral/pkg/log"
 	"github.com/submariner-io/submariner/pkg/packetfilter"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	k8snet "k8s.io/utils/net"
 	"k8s.io/utils/ptr"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
@@ -37,8 +33,7 @@ import (
 
 const (
 	/* Single table named 'submariner' is used for nftables configuration.*/
-	submarinerTable   = "submariner"
-	serializedVersion = "1.0"
+	submarinerTable = "submariner"
 )
 
 var (
@@ -299,201 +294,4 @@ func (p *packetFilter) insertRuleAtPosition(chain string, rule *packetfilter.Rul
 	err := p.nftables.Run(context.TODO(), tx)
 
 	return errors.Wrap(err, "error inserting rule")
-}
-
-func protoToRuleSpec(ruleSpec []string, proto packetfilter.RuleProto, dPort string) []string {
-	switch proto {
-	case packetfilter.RuleProtoUDP:
-		ruleSpec = append(ruleSpec, "ip", "protocol", "udp")
-		if dPort != "" {
-			ruleSpec = append(ruleSpec, "udp", "dport", dPort)
-		}
-	case packetfilter.RuleProtoTCP:
-		ruleSpec = append(ruleSpec, "ip", "protocol", "tcp")
-		if dPort != "" {
-			ruleSpec = append(ruleSpec, "tcp", "dport", dPort)
-		}
-	case packetfilter.RuleProtoICMP:
-		ruleSpec = append(ruleSpec, "ip", "protocol", "icmp")
-	case packetfilter.RuleProtoAll:
-	case packetfilter.RuleProtoUndefined:
-	}
-
-	return ruleSpec
-}
-
-func mssClampToRuleSpec(ruleSpec []string, clampType packetfilter.MssClampType, mssValue string) []string {
-	switch clampType {
-	case packetfilter.UndefinedMSS:
-	case packetfilter.ToPMTU:
-		ruleSpec = append(ruleSpec, "size", "set", "rt", "mtu")
-	case packetfilter.ToValue:
-		ruleSpec = append(ruleSpec, "size", "set", mssValue)
-	}
-
-	return ruleSpec
-}
-
-func setToRuleSpec(ruleSpec []string, srcSetName, destSetName string) []string {
-	if srcSetName != "" {
-		ruleSpec = append(ruleSpec, "ip", "saddr", "@"+srcSetName)
-	}
-
-	if destSetName != "" {
-		ruleSpec = append(ruleSpec, "ip", "daddr", "@"+destSetName)
-	}
-
-	return ruleSpec
-}
-
-func toNftRuleSpec(rule *packetfilter.Rule) string {
-	ruleSpec := protoToRuleSpec([]string{}, rule.Proto, rule.DPort)
-
-	if rule.SrcCIDR != "" {
-		ruleSpec = append(ruleSpec, "ip", "saddr", rule.SrcCIDR)
-	}
-
-	if rule.DestCIDR != "" {
-		ruleSpec = append(ruleSpec, "ip", "daddr", rule.DestCIDR)
-	}
-
-	if rule.MarkValue != "" && rule.Action != packetfilter.RuleActionMark {
-		// syntax for MarkValue '0xc0000': meta mark & 0xc0000 == 0xc0000
-		ruleSpec = append(ruleSpec, "meta", "mark", "&", rule.MarkValue, "==", rule.MarkValue)
-	}
-
-	ruleSpec = setToRuleSpec(ruleSpec, rule.SrcSetName, rule.DestSetName)
-
-	if rule.OutInterface != "" {
-		ruleSpec = append(ruleSpec, "oifname", rule.OutInterface)
-	}
-
-	if rule.InInterface != "" {
-		ruleSpec = append(ruleSpec, "iifname", rule.InInterface)
-	}
-
-	if rule.Action == packetfilter.RuleActionMss {
-		ruleSpec = append(ruleSpec, "tcp", "flags", "syn / syn,rst")
-	}
-
-	ruleSpec = append(ruleSpec, "counter")
-	ruleSpec = append(ruleSpec, ruleActionToStr[rule.Action]...)
-
-	if rule.Action == packetfilter.RuleActionJump {
-		ruleSpec = append(ruleSpec, rule.TargetChain)
-	}
-
-	if rule.SnatCIDR != "" {
-		ruleSpec = append(ruleSpec, "to", rule.SnatCIDR)
-	}
-
-	if rule.DnatCIDR != "" {
-		ruleSpec = append(ruleSpec, "to", rule.DnatCIDR)
-	}
-
-	ruleSpec = mssClampToRuleSpec(ruleSpec, rule.ClampType, rule.MssValue)
-
-	if rule.MarkValue != "" && rule.Action == packetfilter.RuleActionMark {
-		// syntax for MarkValue '0xc0000': set mark | 0xc0000
-		ruleSpec = append(ruleSpec, "set", "mark", "|", rule.MarkValue)
-	}
-
-	str := strings.Join(ruleSpec, " ")
-
-	logger.V(log.TRACE).Infof("toNftRuleSpec: from %q to %q", rule, str)
-
-	return str
-}
-
-func mustWriteString(buf *bytes.Buffer, s string) {
-	_, err := buf.WriteString(s + " ")
-	utilruntime.Must(err)
-}
-
-func mustWriteUint32(buf *bytes.Buffer, i uint32) {
-	mustWriteString(buf, strconv.FormatInt(int64(i), 10))
-}
-
-func SerializeRule(rule *packetfilter.Rule) string {
-	var buf bytes.Buffer
-
-	_, err := buf.WriteString(serializedVersion)
-	utilruntime.Must(err)
-
-	mustWriteUint32(&buf, uint32(rule.Action))
-	mustWriteUint32(&buf, uint32(rule.ClampType))
-	mustWriteUint32(&buf, uint32(rule.Proto))
-
-	mustWriteString(&buf, rule.SrcSetName)
-	mustWriteString(&buf, rule.DestSetName)
-	mustWriteString(&buf, rule.SrcCIDR)
-	mustWriteString(&buf, rule.DestCIDR)
-	mustWriteString(&buf, rule.SnatCIDR)
-	mustWriteString(&buf, rule.DnatCIDR)
-	mustWriteString(&buf, rule.OutInterface)
-	mustWriteString(&buf, rule.InInterface)
-	mustWriteString(&buf, rule.DPort)
-	mustWriteString(&buf, rule.TargetChain)
-	mustWriteString(&buf, rule.MssValue)
-	mustWriteString(&buf, rule.MarkValue)
-
-	str := buf.String()
-
-	logger.V(log.TRACE).Infof("SerializeRule: from %q to %q (%d)", rule, str, len(str))
-
-	return str
-}
-
-func mustReadString(buf *bytes.Buffer) string {
-	b, err := buf.ReadBytes(' ')
-	utilruntime.Must(err)
-
-	return string(b[0 : len(b)-1])
-}
-
-func mustReadUint32(buf *bytes.Buffer) uint32 {
-	s := mustReadString(buf)
-	u, err := strconv.ParseUint(s, 10, 32)
-	utilruntime.Must(err)
-
-	return uint32(u)
-}
-
-func DeserializeRule(s string) (*packetfilter.Rule, error) {
-	rule := &packetfilter.Rule{}
-
-	buf := bytes.NewBufferString(s)
-
-	b := make([]byte, 3)
-
-	_, err := buf.Read(b)
-	if err != nil {
-		return nil, errors.Wrapf(err, "error deserializing rule %q", s)
-	}
-
-	version := string(b)
-	if version != serializedVersion {
-		return nil, errors.Errorf("unable to deserialize rule from %q: invalid version %q", s, version)
-	}
-
-	rule.Action = packetfilter.RuleAction(mustReadUint32(buf))
-	rule.ClampType = packetfilter.MssClampType(mustReadUint32(buf))
-	rule.Proto = packetfilter.RuleProto(mustReadUint32(buf))
-
-	rule.SrcSetName = mustReadString(buf)
-	rule.DestSetName = mustReadString(buf)
-	rule.SrcCIDR = mustReadString(buf)
-	rule.DestCIDR = mustReadString(buf)
-	rule.SnatCIDR = mustReadString(buf)
-	rule.DnatCIDR = mustReadString(buf)
-	rule.OutInterface = mustReadString(buf)
-	rule.InInterface = mustReadString(buf)
-	rule.DPort = mustReadString(buf)
-	rule.TargetChain = mustReadString(buf)
-	rule.MssValue = mustReadString(buf)
-	rule.MarkValue = mustReadString(buf)
-
-	logger.V(log.TRACE).Infof("DeserializeRule: from %q to %q", s, rule)
-
-	return rule, nil
 }
